refactor(wiki): extract upload into its own file, add session tests
Two deslop-flagged code-health items addressed.

1. wiki/write.go was 1,277 LOC (worsened from a 1,121 baseline) mixing
   edit, find/replace, upload, move, and category concerns. The upload
   family (UploadFile, performUpload, uploadFromURL, uploadFromFile,
   uploadFromBytes, parseUploadResponse, parseJSONResponse) is fully
   self-contained and moves cleanly to wiki/upload.go. write.go drops
   to 978 LOC, back below baseline.

   While moving, fold the dead readLocalFile placeholder into
   uploadFromFile directly: readLocalFile always returned an error,
   making the err-check in uploadFromFile structurally always-true.
   The behavior is unchanged (local-path uploads still refused with the
   same message); staticcheck no longer flags the always-true branch.

2. wiki/session.go had no unit tests. SessionSnapshot and RestoreSession
   silently round-trip cookies and auth state across CLI invocations —
   exactly the surface where bugs fail invisibly. Add session_test.go
   with seven cases: empty snapshot, snapshot round-trip, expired-token
   restore (cookies transfer but loggedIn stays false), nil-jar branch,
   invalid BaseURL on both snapshot and restore, and empty-snapshot
   restore safety.
